fix(format): make printf specifiers portable across host and RV32 (#93)
The sketch warning gate (next commit) compiles example sketches under
-Wall -Wextra -Werror -Wswitch with the RV32 cross-compiler, where
uint32_t is long unsigned int. ~50 sites passed uint32_t or unsigned
long values to %u converters -- a -Wformat mismatch on RV32 that the
host compiler (macOS, where uint32_t is unsigned int) never flagged.
The code compiled everywhere but was non-portable.

Fixed by wrapping each such argument in static_cast<unsigned>(value)
paired with %u, the tree's existing convention (e.g.
static_cast<unsigned>(host_->nodeCount()) in TracerShell). unsigned is
unsigned int on every target, so the cast is a no-op on the host and a
narrowing to match %u on RV32. No struct field types or public APIs
changed.

Sites: src/testbed/TracerShell.h (emitTrace_, emitBackoffTrace_,
emitHostLine_, emitNodeLine_ -- seq/ts, the host/link/decoder stats
block, the degraded-lane counters, the transport getters, the exchange
counters, the breaker counters), src/SimpleHostMetrics.h (headerText,
nodeRowText), examples/XiaoSniffer/XiaoSniffer.ino (emitEpoch,
emitFrame, emitStats). Already-correct %lu + static_cast<unsigned long>
sites left untouched.

Also removed an unused `using VerbResult` alias in XiaoHostTracer.ino
flagged as -Wunused-local-typedefs; the sibling alias that IS used
(VerbResult::kQuit) is kept.

make -C tests test (219 tests) and make -C extras/desktop both green.

+++ src/SimpleHostMetrics.h
@@ -189,7 +189,7 @@ class HostStatusPanel {
       if (interval == 0) {
         snprintf(buf, len, "---ms");
       } else {
-        snprintf(buf, len, "%ums", interval);
+        snprintf(buf, len, "%ums", static_cast<unsigned>(interval));
       }
     }
   }
@@ -209,7 +209,7 @@ class HostStatusPanel {
         ? nodeErrors_[nodeIndex].countInLastMs(nowMs, kErrorWindowMs)
         : 0;
     snprintf(buf, len, "UA%u:%s %s %2uerr", address,
-             stateTag ? stateTag : "---", lat, recentErrs);
+             stateTag ? stateTag : "---", lat, static_cast<unsigned>(recentErrs));
   }
 
  private:
